updated test

+++ TestApplication/TestApplication/TestApplication.py
@@ -5,7 +5,7 @@ age = int(input('Please enter your age: '))
 factor = 2
 finalAge = age + factor
 multAge = age * factor
-divAge = age / factor
+divAge = float(age / factor)
 
 print('In', factor, 'years you will be', finalAge, 'years old', userName)
 print('Your age multiplied by', factor, 'is', multAge)
